% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denmark.R
\name{denmark}
\alias{denmark}
\title{Dataset of boundaries for Denmark}
\usage{
denmark()
}
\value{
A \code{polyclid_polygon} vector
}
\description{
This function returns a polygon vector encoding the various geographical
areas of Denmark at a high resolution.
}
\examples{

dk <- denmark()

plot(dk)

}
