% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyline_monotone.R
\name{is_monotone}
\alias{is_monotone}
\alias{is_weakly_monotone}
\title{Polyline monotonicity}
\usage{
is_monotone(x, dim = "both")

is_weakly_monotone(x, dim = "x")
}
\arguments{
\item{x}{A \code{polyclid_polyline} vector}

\item{dim}{Either \code{"x"} or \code{"y"} to indicate in which dimension to test
monotonicity in}
}
\value{
A logical vector with the same length as \code{x}
}
\description{
Monotonicity describes the property that values along a line is ever
increasing or decreasing, i.e. \code{f(x) > f(x + 1) || f(x) < f(x + 1)}. The weak
version allows equality as well, i.e. \code{f(x) >= f(x + 1) || f(x) <= f(x + 1)}.
}
\examples{
sine <- polyline(
  seq(0, 2*pi, length.out = 20),
  sin(seq(0, 2*pi, length.out = 20))
)
is_monotone(sine, "x")
is_monotone(sine, "y")

poly <- polyline(
  c(0, 1, 1, 2),
  c(0, 1, 2, 3)
)

is_monotone(sine, "x")
is_weakly_monotone(sine, "x")

}
