% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_topology.R
\name{polygon_topology}
\alias{polygon_topology}
\alias{is_convex}
\alias{is_simple}
\alias{is_relatively_simple}
\title{Polygon topology}
\usage{
is_convex(x)

is_simple(x)

is_relatively_simple(x)
}
\arguments{
\item{x}{A \code{polyclid_polygon} vector}
}
\value{
a logical vector of the same length as \code{x}
}
\description{
These functions checks certain aspects of a polygons topology. They all
operate only on the outer boundary of a polygon and thus ignores all holes it
may have. A convex polygon is one whose boundary exclusively have the same
turn direction as it is traversed. A simply polygon is a polygon with no
intersections between its edges except at the end points of neighboring
edges). A relatively simple polygon is like a simple polygon except it
allows a shared end point between non-neighboring edges, with the requirement
that it remains orientable (it must not "cross over" at a vertex). A
relatively simple polygon is always simple. Relative simplicity is one of the
requirements of rings in a valid polygon.
}
\examples{

square <- polygon(
  c(0, 10, 10, 0),
  c(0, 0, 10, 10)
)
cross <- polygon(
  c(0, 10, 10, 0),
  c(0, 10, 0, 10)
)
butterfly <- polygon(
  c(0, 5, 10, 10, 5, 0),
  c(0, 5, 0, 10, 5, 10)
)
cross2 <- polygon(
  c(0, 5, 10, 10, 5, 0),
  c(0, 5, 10, 0, 5, 10)
)
is_convex(square)
is_convex(cross)
is_convex(butterfly)
is_convex(cross2)

is_simple(square)
is_simple(cross)
is_simple(butterfly)
is_simple(cross2)

is_relatively_simple(square)
is_relatively_simple(cross)
is_relatively_simple(butterfly)
is_relatively_simple(cross2)
}
