% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyclid-package.R
\docType{package}
\name{polyclid-package}
\alias{polyclid}
\alias{polyclid-package}
\title{polyclid: Polygon Support for Euclid}
\description{
Provides a polygon and polyline vector type for the 'euclid' ecosystem, along with several algorithms specific to these geometry types.
}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

}
\keyword{internal}
